working version

+++ src/parse_data.py
@@ -9,10 +9,10 @@ def parse_planet_data(filename):
         all_planets = []
         for planet in csv_reader:
             name = planet[0]
-            mass = float(planet[1])
-            position = np.array([float(planet[2]), float(planet[3])])
-            velocity = np.array([float(planet[4]), float(planet[5])])
-            radius = float(planet[6])
+            mass = np.float64(planet[1])
+            position = np.array([np.float64(planet[2]), np.float64(planet[3])])
+            velocity = np.array([np.float64(planet[4]), np.float64(planet[5])])
+            radius = np.float64(planet[6])
             colour = str(planet[7])
             p = Planet(name, mass, position, velocity, radius, colour)
             all_planets.append(p)
